Refactor ErrorBoundary into ErrorFallback
